Bundle Maestro for use in browser

+++ app/writer/task/HtmlTaskWriter.js
@@ -6,8 +6,10 @@ const path = require('path');
 const getImageFileDimensions = require('image-size');
 const nunjucks = require('../../model/nunjucksEnvironment');
 const consoleHelper = require('../../helpers/consoleHelper');
+const envHelper = require('../../helpers/envHelper');
 const TaskWriter = require('./TaskWriter');
 const TextTransform = require('../TextTransform');
+const settings = require('../../settings');
 
 module.exports = class HtmlTaskWriter extends TaskWriter {
 
@@ -35,7 +37,7 @@ module.exports = class HtmlTaskWriter extends TaskWriter {
 			// copy image from ./images to ./build
 			// Do this asynchronously...no need to wait
 			// Also, super lazy: if the image already exists don't copy it again
-			if (!fs.existsSync(imageBuildPath)) {
+			if (envHelper.isNode && !fs.existsSync(imageBuildPath)) {
 				fs.copyFile(imageSrcPath, imageBuildPath, (err) => {
 					if (err) {
 						// for now don't throw errors on this. Allow build to finish
@@ -46,7 +48,7 @@ module.exports = class HtmlTaskWriter extends TaskWriter {
 			}
 
 			const image = nunjucks.render('image.html', {
-				path: imageMeta.path,
+				path: settings.htmlImagePrefix + imageMeta.path,
 				width: imageSize.width,
 				height: imageSize.height
 			});
